Refactored module structure
