+++ crypto-benchmarks.rs/Cargo.toml
@@ -4,6 +4,7 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+bitvec = "1.0.1"
 blst = "0.3.13"
 criterion = "0.5.1"
 pallas = "0.32.0"
