+++ crypto-benchmarks.rs/benches/vrf_bench.rs
@@ -1,4 +1,5 @@
 use criterion::{criterion_group, criterion_main, Criterion};
+
 use leios_crypto_benchmarks::vrf::*;
 
 fn benchmark_prove(c: &mut Criterion) {
