+++ crypto-benchmarks.rs/src/bls_util.rs
