+++ crypto-benchmarks.rs/src/lib.rs
@@ -1,5 +1,8 @@
-pub mod api;
-pub mod bls;
+pub mod bls_util;
+pub mod bls_vote;
+pub mod cert;
+pub mod key;
+pub mod primitive;
 pub mod sortition;
 pub mod util;
 pub mod vote;
